Extract callable initX() function from ies implementation
The iterated ensemble smoother contains the functionality calculate initialize
an X matrix inside the implementation. The main content of this PR is to extract
that functionality into a separate function which can be called as a single
update, without considering iteration.

In addition to this there are many refactorings in the code:

 - The amount of logging and debug messages is severely reduced.
 - The code to manage state like matrices E, A and W is simplified.
 - Some fuctionality has been extracted into small functions.